requirements.txt
streamlit
pandas
chromadb
sentence-transformers
openai
python-dotenv
pymupdf
tqdm